Fixed serialization. Refactoring code

+++ BusinessSharTests/Core/WarehouseTests.cs
@@ -19,7 +19,7 @@ namespace BusinessSharkTests.Core
         [SetUp]
         public void SetUp()
         {
-            _itemDef = new ItemDefinition((int)Enums.ItemType.Wood, "Wood", 1, 1, 0.1f, 0.1f, 0.1f, 0.1f, 0.1f, 0.1f,
+            _itemDef = new ItemDefinition(Enums.ItemType.Wood, "Wood", 1, 1, 0.1f, 0.1f, 0.1f, 0.1f, 0.1f, 0.1f,
                 0.1f);
             _fromDivision = new Warehouse(1, "TestWarehouseFrom", new Location(), int.MaxValue);
             _toDivision = new Warehouse(2, "TestWarehouseTo", new Location(1, 1), int.MaxValue);
